# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Conjur API
  description: OpenAPI specification for Conjur - compatible with CyberArk Conjur
  version: "5.0.0"
  license:
    name: Apache-2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Basic authentication with username and API key
    bearerAuth:
      type: http
      scheme: bearer
      description: Conjur access token
  schemas:
...
    AuthenticatorsResponse:
      type: object
      properties:
        installed:
          type: array
          items:
            type: string
        enabled:
          type: array
          items:
            type: string
        configured:
          type: array
          items:
            type: string
        status:
          type: string
...
  parameters:
    Account:
      name: account
      in: path
      required: true
      schema:
        type: string
      description: Organization account name
...
    Kind:
      name: kind
      in: path
      required: true
      schema:
        type: string
      description: Resource kind
    Identifier:
      name: identifier
      in: path
      required: true
      schema:
        type: string
      description: URL-encoded resource identifier
...
    ServiceId:
      name: service_id
      in: path
      required: true
      schema:
        type: string
      description: Authenticator service ID
    Authenticator:
      name: authenticator
      in: path
      required: true
      schema:
        type: string
      description: Authenticator type
paths:
...
  /{authenticator}/{account}/status:
    get:
      operationId: getAuthenticatorStatus
      summary: Get authenticator status
      tags: [status]
      security: []
      parameters:
        - $ref: "#/components/parameters/Authenticator"
        - $ref: "#/components/parameters/Account"
      responses:
        "200":
          description: Authenticator status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticatorsResponse"
        "501":
          description: Authenticator not enabled
  /{authenticator}/{service_id}/{account}/status:
    get:
      operationId: getAuthenticatorServiceStatus
      summary: Get authenticator service status
      tags: [status]
      security: []
      parameters:
        - $ref: "#/components/parameters/Authenticator"
        - $ref: "#/components/parameters/ServiceId"
        - $ref: "#/components/parameters/Account"
      responses:
        "200":
          description: Authenticator status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticatorsResponse"
        "501":
          description: Authenticator not enabled

  # ========== ACCOUNTS ==========
...
  /secrets/{account}/{kind}/{identifier}:
    get:
      operationId: getSecret
      summary: Fetch a secret value
      tags: [secrets]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Account"
        - $ref: "#/components/parameters/Kind"
        - $ref: "#/components/parameters/Identifier"
        - name: version
          in: query
          schema:
            type: integer
          description: Secret version (default latest)
      responses:
        "200":
          description: Secret value
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "403":
          description: Forbidden
        "404":
          description: Secret not found

    post:
      operationId: createSecret
      summary: Create or update a secret value, or expire it
      tags: [secrets]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Account"
        - $ref: "#/components/parameters/Kind"
        - $ref: "#/components/parameters/Identifier"
        - name: expirations
          in: query
          schema:
            type: string
          description: If present, expire the secret instead of setting value
      requestBody:
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: Secret created/expired
        "403":
          description: Forbidden
        "422":
          description: Invalid kind for expiration
  /secrets:
    get:
      operationId: batchGetSecrets
      summary: Fetch multiple secrets
      tags: [secrets]
      security:
        - bearerAuth: []
      parameters:
        - name: variable_ids
          in: query
          required: true
          schema:
            type: string
          description: Comma-separated list of variable IDs
      responses:
        "200":
          description: Map of secret values
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        "403":
          description: Forbidden
        "404":
          description: One or more secrets not found
  /secrets/{account}/values:
    post:
      operationId: batchUpdateSecrets
      summary: Update multiple secrets
      tags: [secrets]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Account"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                secrets:
                  type: object
                  additionalProperties:
                    type: string
      responses:
        "201":
          description: Secrets updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: array
                    items:
                      type: string
                  errors:
                    type: object
                    additionalProperties:
                      type: string
        "207":
          description: Partial success

  # ========== POLICIES ==========
...